+++ src/promptflow/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Release History
 
+## 1.3.0 (Upcoming)
+
+### Features Added
+- [SDK/CLI] Support `pfazure run cancel` to cancel a run on Azure AI.
+
+### Improvements
+- Change `ruamel.yaml` lower bound to 0.17.10.
+- [SDK/CLI] Improve `pfazure run download` to handle large run data files.
+
 ## 1.2.0 (2023.12.14)
 
 ### Features Added